some fix associated with system exit when shader compilation filed and still tuning DoF demo.

+++ GLFWApp.cpp
@@ -110,11 +110,11 @@ void GLFWApp::setPinMouse(bool b) {
     this->isPinMouse = b;
 }
 
-int GLFWApp::getDeltaMouseX() {
+float GLFWApp::getDeltaMouseX() {
     return this->deltaMouseX;
 }
 
-int GLFWApp::getDeltaMouseY() {
+float GLFWApp::getDeltaMouseY() {
     return this->deltaMouseY;
 }
